missed a namespace move
